+++ tests/test_add_option_kwargs.py
@@ -3,46 +3,6 @@ from loguru import logger
 import io
 
 
-def test_function_with_kwargs():
-    out = []
-
-    def function(message, kw2, kw1):
-        out.append(message + kw1 + "a" + kw2)
-
-    logger.add(function, format="{message}", kw1="1", kw2="2")
-    logger.debug("msg")
-    assert out == ["msg\n1a2"]
-
-
-def test_class_with_kwargs():
-    out = []
-
-    class Writer:
-        def __init__(self, kw2, kw1):
-            self.end = kw1 + "b" + kw2
-
-        def write(self, m):
-            out.append(m + self.end)
-
-    logger.add(Writer, format="{message}", kw1="1", kw2="2")
-    logger.debug("msg")
-    assert out == ["msg\n1b2"]
-
-
-def test_file_object_with_kwargs():
-    class Writer:
-        def __init__(self):
-            self.out = ""
-
-        def write(self, m, kw2, kw1):
-            self.out += m + kw1 + "c" + kw2
-
-    writer = Writer()
-    logger.add(writer, format="{message}", kw1="1", kw2="2")
-    logger.debug("msg")
-    assert writer.out == "msg\n1c2"
-
-
 def test_file_mode_a(tmpdir):
     file = tmpdir.join("test.log")
     file.write("base\n")
@@ -69,20 +29,11 @@ def test_file_buffering(tmpdir):
 
 
 def test_invalid_function_kwargs():
-    def function(message, a="Y"):
-        pass
-
-    logger.add(function, b="X", catch=False)
-    with pytest.raises(TypeError):
-        logger.debug("Nope")
-
-
-def test_invalid_class_kwargs():
-    class Writer:
+    def function(message):
         pass
 
     with pytest.raises(TypeError):
-        logger.add(Writer, keyword=123)
+        logger.add(function, b="X")
 
 
 def test_invalid_file_object_kwargs():
@@ -94,9 +45,9 @@ def test_invalid_file_object_kwargs():
             pass
 
     writer = Writer()
-    logger.add(writer, format="{message}", kw1="1", kw2="2", catch=False)
+
     with pytest.raises(TypeError):
-        logger.debug("msg")
+        logger.add(writer, format="{message}", kw1="1", kw2="2")
 
 
 def test_invalid_file_kwargs():